change integer to bigdecimal

+++ src/main/java/br/unioeste/padaria/receita/model/entity/Receita.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -23,7 +24,7 @@ public class Receita {
 
     String nomeReceita;
 
-    Integer rendimento;
+    BigDecimal rendimento;
 
     Integer tempoPreparo;
